Split daemon in bin and lib
